Refactor import paths to use the new state directory structure for AppState
