+++ ui/lib/components/messages/message_input.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 // Project imports:
-import 'package:ui/app_state.dart';
+import 'package:ui/state/app_state.dart';
 import 'package:ui/helpers/http_helper.dart';
 import 'package:ui/state/notification_state.dart';
 
@@ -25,7 +25,8 @@ class _MessageInputState extends State<MessageInput> {
 
   void sendMessage() async {
     final appState = Provider.of<AppState>(context, listen: false);
-    final notificationState = Provider.of<NotificationState>(context);
+    final notificationState =
+        Provider.of<NotificationState>(context, listen: false);
 
     final String userMessage = textController.text.trim();
     if (userMessage.isEmpty) {
