+++ ui/test/pages/login_page_test.dart
@@ -8,7 +8,7 @@ import 'package:mockito/annotations.dart';
 import 'package:provider/provider.dart';
 
 // Project imports:
-import 'package:ui/app_state.dart';
+import 'package:ui/state/app_state.dart';
 import 'package:ui/helpers/http_helper.dart';
 import 'package:ui/pages/login_page.dart';
 import 'package:ui/state/notification_state.dart';
